backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
gevent==25.9.1
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.7.0
zope.event==6.1
zope.interface==8.1.1
